Revert "[Turbopack] Production Chunking (#75010)"
This reverts commit 169829a156761bf83b1ff3a0cc6deb12b716a7f6.
