+++ crates/next-core/src/next_server/context.rs
@@ -14,7 +14,7 @@ use turbopack::{
     transition::Transition,
 };
 use turbopack_core::{
-    chunk::{module_id_strategies::ModuleIdStrategy, ChunkingConfig, MinifyType},
+    chunk::{module_id_strategies::ModuleIdStrategy, MinifyType},
     compile_time_info::{
         CompileTimeDefineValue, CompileTimeDefines, CompileTimeInfo, DefineableNameSegment,
         FreeVarReferences,
@@ -1010,10 +1010,7 @@ pub async fn get_server_chunking_context_with_client_assets(
 
     if next_mode.is_development() {
         builder = builder.use_file_source_map_uris();
-    } else {
-        builder = builder.ecmascript_chunking_config(ChunkingConfig {})
     }
-
     Ok(builder.build())
 }
 
@@ -1051,8 +1048,6 @@ pub async fn get_server_chunking_context(
 
     if next_mode.is_development() {
         builder = builder.use_file_source_map_uris()
-    } else {
-        builder = builder.ecmascript_chunking_config(ChunkingConfig {})
     }
 
     Ok(builder.build())
