+++ turbopack/crates/turbopack-core/Cargo.toml
@@ -26,7 +26,6 @@ patricia_tree = "0.5.5"
 petgraph = { workspace = true, features = ["serde-1"] }
 roaring = { version = "0.10.10", features = ["serde"] }
 ref-cast = "1.0.20"
-rustc-hash ={ workspace = true }
 regex = { workspace = true }
 serde = { workspace = true, features = ["rc"] }
 serde_bytes = { workspace = true }
@@ -34,7 +33,6 @@ serde_json = { workspace = true, features = ["preserve_order"] }
 sourcemap = { workspace = true }
 swc_core = { workspace = true, features = ["ecma_preset_env", "common"] }
 tracing = { workspace = true }
-turbo-prehash = { workspace = true }
 turbo-rcstr = { workspace = true }
 turbo-tasks = { workspace = true }
 turbo-tasks-env = { workspace = true }
