+++ turbopack/crates/turbopack-nodejs/Cargo.toml
@@ -22,7 +22,6 @@ workspace = true
 anyhow = { workspace = true }
 indexmap = { workspace = true }
 indoc = { workspace = true }
-rustc-hash = { workspace = true }
 serde = { workspace = true }
 tracing = { workspace = true }
 urlencoding = { workspace = true }
